api/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.3
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
librabbitmq==2.0.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.39
pycodestyle==2.11.1
pydantic==2.4.2
pydantic-core==2.10.1
PyJWT==2.8.0
pymongo==4.5.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
schema==0.7.5
six==1.16.0
tomli==2.0.1
typing-extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
werkzeug==3.0.0
zipp==3.17.0
